+++ app/src/commonMain/kotlin/model/Point.kt
@@ -20,6 +20,6 @@
 package model
 
 data class Point(
-    val x: Short,
-    val y: Short
+    val x: Int,
+    val y: Int
 )
